TST/REF: split/collect large tests (#39789)
