+++ pandas/tests/frame/methods/test_set_index.py
@@ -1,3 +1,7 @@
+"""
+See also: test_reindex.py:TestReindexSetIndex
+"""
+
 from datetime import datetime, timedelta
 
 import numpy as np
