+++ pandas/tests/frame/test_nonunique_indexes.py
@@ -33,6 +33,7 @@ class TestDataFrameNonuniqueIndexes:
         expected = DataFrame([[1, 1, 1, 5], [1, 1, 2, 5], [2, 1, 3, 5]], columns=idx)
         check(df, expected)
 
+    def test_insert_with_duplicate_columns(self):
         # insert
         df = DataFrame(
             [[1, 1, 1, 5], [1, 1, 2, 5], [2, 1, 3, 5]],
@@ -119,6 +120,7 @@ class TestDataFrameNonuniqueIndexes:
         )
         tm.assert_frame_equal(df, expected)
 
+    def test_dup_across_dtypes(self):
         # dup across dtypes
         df = DataFrame(
             [[1, 1, 1.0, 5], [1, 1, 2.0, 5], [2, 1, 3.0, 5]],
@@ -155,12 +157,14 @@ class TestDataFrameNonuniqueIndexes:
         )
         check(df, expected)
 
+    def test_values_with_duplicate_columns(self):
         # values
         df = DataFrame([[1, 2.5], [3, 4.5]], index=[1, 2], columns=["x", "x"])
         result = df.values
         expected = np.array([[1, 2.5], [3, 4.5]])
         assert (result == expected).all().all()
 
+    def test_rename_with_duplicate_columns(self):
         # rename, GH 4403
         df4 = DataFrame(
             {"RT": [0.0454], "TClose": [22.02], "TExg": [0.0422]},
@@ -201,6 +205,8 @@ class TestDataFrameNonuniqueIndexes:
         ).set_index(["STK_ID", "RPT_Date"], drop=False)
         tm.assert_frame_equal(result, expected)
 
+    def test_reindex_with_duplicate_columns(self):
+
         # reindex is invalid!
         df = DataFrame(
             [[1, 5, 7.0], [1, 5, 7.0], [1, 5, 7.0]], columns=["bar", "a", "a"]
@@ -210,6 +216,8 @@ class TestDataFrameNonuniqueIndexes:
             df.reindex(columns=["bar"])
         with pytest.raises(ValueError, match=msg):
             df.reindex(columns=["bar", "foo"])
+
+    def test_drop_with_duplicate_columns(self):
 
         # drop
         df = DataFrame(
@@ -221,6 +229,7 @@ class TestDataFrameNonuniqueIndexes:
         result = df.drop("a", axis=1)
         check(result, expected)
 
+    def test_describe_with_duplicate_columns(self):
         # describe
         df = DataFrame(
             [[1, 1, 1], [2, 2, 2], [3, 3, 3]],
@@ -232,6 +241,7 @@ class TestDataFrameNonuniqueIndexes:
         expected = pd.concat([s, s, s], keys=df.columns, axis=1)
         check(result, expected)
 
+    def test_column_dups_indexes(self):
         # check column dups with index equal and not equal to df's index
         df = DataFrame(
             np.random.randn(5, 3),
@@ -248,6 +258,8 @@ class TestDataFrameNonuniqueIndexes:
             this_df["A"] = index
             check(this_df, expected_df)
 
+    def test_arithmetic_with_dups(self):
+
         # operations
         for op in ["__add__", "__mul__", "__sub__", "__truediv__"]:
             df = DataFrame({"A": np.arange(10), "B": np.random.rand(10)})
@@ -257,6 +269,7 @@ class TestDataFrameNonuniqueIndexes:
             result = getattr(df, op)(df)
             check(result, expected)
 
+    def test_changing_dtypes_with_duplicate_columns(self):
         # multiple assignments that change dtypes
         # the location indexer is a slice
         # GH 6120
@@ -272,7 +285,7 @@ class TestDataFrameNonuniqueIndexes:
         df["that"] = 1
         check(df, expected)
 
-    def test_column_dups2(self):
+    def test_column_dups_drop(self):
 
         # drop buggy GH 6240
         df = DataFrame(
@@ -289,6 +302,7 @@ class TestDataFrameNonuniqueIndexes:
         result = df2.drop("C", axis=1)
         tm.assert_frame_equal(result, expected)
 
+    def test_column_dups_dropna(self):
         # dropna
         df = DataFrame(
             {
@@ -310,43 +324,6 @@ class TestDataFrameNonuniqueIndexes:
         result = df.dropna(subset=["A", "C"], how="all")
         tm.assert_frame_equal(result, expected)
 
-    def test_getitem_boolean_series_with_duplicate_columns(self):
-        # boolean indexing
-        # GH 4879
-        dups = ["A", "A", "C", "D"]
-        df = DataFrame(
-            np.arange(12).reshape(3, 4), columns=["A", "B", "C", "D"], dtype="float64"
-        )
-        expected = df[df.C > 6]
-        expected.columns = dups
-        df = DataFrame(np.arange(12).reshape(3, 4), columns=dups, dtype="float64")
-        result = df[df.C > 6]
-        check(result, expected)
-
-    def test_getitem_boolean_frame_with_duplicate_columns(self):
-        dups = ["A", "A", "C", "D"]
-
-        # where
-        df = DataFrame(
-            np.arange(12).reshape(3, 4), columns=["A", "B", "C", "D"], dtype="float64"
-        )
-        # `df > 6` is a DataFrame with the same shape+alignment as df
-        expected = df[df > 6]
-        expected.columns = dups
-        df = DataFrame(np.arange(12).reshape(3, 4), columns=dups, dtype="float64")
-        result = df[df > 6]
-        check(result, expected)
-
-    def test_getitem_boolean_frame_unaligned_with_duplicate_columns(self):
-        # `df.A > 6` is a DataFrame with a different shape from df
-        dups = ["A", "A", "C", "D"]
-
-        # boolean with the duplicate raises
-        df = DataFrame(np.arange(12).reshape(3, 4), columns=dups, dtype="float64")
-        msg = "cannot reindex from a duplicate axis"
-        with pytest.raises(ValueError, match=msg):
-            df[df.A > 6]
-
     def test_column_dups_indexing(self):
 
         # dup aligning operations should work
@@ -357,6 +334,7 @@ class TestDataFrameNonuniqueIndexes:
         result = df1.sub(df2)
         tm.assert_frame_equal(result, expected)
 
+    def test_dup_columns_comparisons(self):
         # equality
         df1 = DataFrame([[1, 2], [2, np.nan], [3, 4], [4, 4]], columns=["A", "B"])
         df2 = DataFrame([[0, 1], [2, 4], [2, np.nan], [4, 5]], columns=["A", "A"])
@@ -374,6 +352,7 @@ class TestDataFrameNonuniqueIndexes:
         )
         tm.assert_frame_equal(result, expected)
 
+    def test_mixed_column_selection(self):
         # mixed column selection
         # GH 5639
         dfbool = DataFrame(
@@ -387,6 +366,7 @@ class TestDataFrameNonuniqueIndexes:
         result = dfbool[["one", "three", "one"]]
         check(result, expected)
 
+    def test_multi_axis_dups(self):
         # multi-axis dups
         # GH 6121
         df = DataFrame(
@@ -422,6 +402,7 @@ class TestDataFrameNonuniqueIndexes:
         expected = DataFrame([[1, 2, 3]], columns=["b", "a", "a.1"])
         tm.assert_frame_equal(df, expected)
 
+    def test_columns_with_dup_index(self):
         # with a dup index
         df = DataFrame([[1, 2]], columns=["a", "a"])
         df.columns = ["b", "b"]
@@ -429,6 +410,7 @@ class TestDataFrameNonuniqueIndexes:
         expected = DataFrame([[1, 2]], columns=["b", "b"])
         tm.assert_frame_equal(df, expected)
 
+    def test_multi_dtype(self):
         # multi-dtype
         df = DataFrame(
             [[1, 2, 1.0, 2.0, 3.0, "foo", "bar"]],
@@ -441,12 +423,14 @@ class TestDataFrameNonuniqueIndexes:
         )
         tm.assert_frame_equal(df, expected)
 
+    def test_multi_dtype2(self):
         df = DataFrame([[1, 2, "foo", "bar"]], columns=["a", "a", "a", "a"])
         df.columns = ["a", "a.1", "a.2", "a.3"]
         str(df)
         expected = DataFrame([[1, 2, "foo", "bar"]], columns=["a", "a.1", "a.2", "a.3"])
         tm.assert_frame_equal(df, expected)
 
+    def test_dups_across_blocks(self):
         # dups across blocks
         df_float = DataFrame(np.random.randn(10, 3), dtype="float64")
         df_int = DataFrame(np.random.randn(10, 3), dtype="int64")
@@ -464,6 +448,7 @@ class TestDataFrameNonuniqueIndexes:
         for i in range(len(df.columns)):
             df.iloc[:, i]
 
+    def test_dup_columns_across_dtype(self):
         # dup columns across dtype GH 2079/2194
         vals = [[1, -1, 2.0], [2, -2, 3.0]]
         rs = DataFrame(vals, columns=["A", "A", "B"])
@@ -486,36 +471,3 @@ class TestDataFrameNonuniqueIndexes:
 
         df.iloc[:, 0] = 3
         tm.assert_series_equal(df.iloc[:, 1], expected)
-
-    @pytest.mark.parametrize(
-        "data1,data2,expected_data",
-        (
-            (
-                [[1, 2], [3, 4]],
-                [[0.5, 6], [7, 8]],
-                [[np.nan, 3.0], [np.nan, 4.0], [np.nan, 7.0], [6.0, 8.0]],
-            ),
-            (
-                [[1, 2], [3, 4]],
-                [[5, 6], [7, 8]],
-                [[np.nan, 3.0], [np.nan, 4.0], [5, 7], [6, 8]],
-            ),
-        ),
-    )
-    def test_masking_duplicate_columns_mixed_dtypes(
-        self,
-        data1,
-        data2,
-        expected_data,
-    ):
-        # GH31954
-
-        df1 = DataFrame(np.array(data1))
-        df2 = DataFrame(np.array(data2))
-        df = pd.concat([df1, df2], axis=1)
-
-        result = df[df > 2]
-        expected = DataFrame(
-            {i: np.array(col) for i, col in enumerate(expected_data)}
-        ).rename(columns={2: 0, 3: 1})
-        tm.assert_frame_equal(result, expected)
